amulet_renderer/requirements.txt
wxPython
numpy==1.15.2
pyopengl
pillow
git+git://github.com/Amulet-Team/Amulet-NBT.git
git+git://github.com/gentlegiantJGC/PyMCTranslate.git
git+git://github.com/Amulet-Team/Amulet-Core.git